Minor refactor of rules
