+++ docs/0151.md
@@ -14,20 +14,20 @@ to complete. Instead of blocking the client, these operations should return a
 `202 Accepted` response with an `Operation` resource that can be used to track
 progress and retrieve the result.
 
-## aep-151-no-200-success
+## aep-151-200-only-success
 
 **Rule**: Long-running operations must return 202 as the only success status
-code, not 200
+code
 
-This rule ensures that operations returning `202 Accepted` do not also define
-`200 OK` as a success response, as mandated in [AEP-151][].
+This rule ensures that operations returning `202 Accepted` do not also define a
+`200`, `201`, or `204` response, as mandated in [AEP-151][].
 
 ### Details
 
 This rule looks for operations that have a `202` response and complains if they
-also have a `200` response defined. Long-running operations must be consistent
-in their behavior - they should either always return `202` with an `Operation`,
-or never do so.
+also have a `200`, `201`, or `204` response defined. Long-running operations
+must be consistent in their behavior - they should either always return `202`
+with an `Operation`, or never do so.
 
 ### Examples
 
@@ -86,137 +86,7 @@ overrides:
       aep-151-no-200-success: 'off'
 ```
 
-## aep-151-no-201-success
-
-**Rule**: Long-running operations must return 202 as the only success status
-code, not 201
-
-This rule ensures that operations returning `202 Accepted` do not also define
-`201 Created` as a success response, as mandated in [AEP-151][].
-
-### Details
-
-This rule looks for operations that have a `202` response and complains if they
-also have a `201` response defined. Long-running operations should not mix
-immediate success responses with asynchronous responses.
-
-### Examples
-
-**Incorrect** code for this rule:
-
-```yaml
-paths:
-  '/books':
-    post:
-      summary: Create a book
-      responses:
-        '201':
-          description: Book created immediately
-          content:
-            'application/json':
-              schema:
-                $ref: '#/components/schemas/Book'
-        '202':
-          description: Book creation accepted
-          content:
-            'application/json':
-              schema:
-                $ref: '#/components/schemas/Operation'
-```
-
-**Correct** code for this rule:
-
-```yaml
-paths:
-  '/books':
-    post:
-      summary: Create a book
-      responses:
-        '202':
-          description: Book creation accepted
-          content:
-            'application/json':
-              schema:
-                $ref: '#/components/schemas/Operation'
-```
-
-### Disabling
-
-If you need to violate this rule for a specific operation, add an "override" to
-the Spectral rule file for the specific file and fragment.
-
-```yaml
-overrides:
-  - files:
-      - 'openapi.json#/paths/books/post/responses'
-    rules:
-      aep-151-no-201-success: 'off'
-```
-
-## aep-151-no-204-success
-
-**Rule**: Long-running operations must return 202 as the only success status
-code, not 204
-
-This rule ensures that operations returning `202 Accepted` do not also define
-`204 No Content` as a success response, as mandated in [AEP-151][].
-
-### Details
-
-This rule looks for operations that have a `202` response and complains if they
-also have a `204` response defined. Long-running operations should not mix
-immediate success responses with asynchronous responses.
-
-### Examples
-
-**Incorrect** code for this rule:
-
-```yaml
-paths:
-  '/books/{id}':
-    delete:
-      summary: Delete a book
-      responses:
-        '204':
-          description: Book deleted immediately
-        '202':
-          description: Book deletion accepted
-          content:
-            'application/json':
-              schema:
-                $ref: '#/components/schemas/Operation'
-```
-
-**Correct** code for this rule:
-
-```yaml
-paths:
-  '/books/{id}':
-    delete:
-      summary: Delete a book
-      responses:
-        '202':
-          description: Book deletion accepted
-          content:
-            'application/json':
-              schema:
-                $ref: '#/components/schemas/Operation'
-```
-
-### Disabling
-
-If you need to violate this rule for a specific operation, add an "override" to
-the Spectral rule file for the specific file and fragment.
-
-```yaml
-overrides:
-  - files:
-      - 'openapi.json#/paths/books/{id}/delete/responses'
-    rules:
-      aep-151-no-204-success: 'off'
-```
-
-## aep-151-202-content-required
+## aep-151-202-schema-required
 
 **Rule**: 202 response must define an application/json response body with
 Operation schema
@@ -227,7 +97,8 @@ with an `application/json` response body, as mandated in [AEP-151][].
 ### Details
 
 This rule looks for `202` responses and complains if they do not have a
-`content` field with an `application/json` media type defined.
+`content` field with an `application/json` media type defined that contains a
+response body schema.
 
 ### Examples
 
@@ -274,81 +145,6 @@ overrides:
 ```
 
 ## aep-151-operation-schema
-
-**Rule**: 202 response must define an application/json response body with
-Operation schema
-
-This rule ensures that the `202` response schema contains the required
-properties for an `Operation` resource, as mandated in [AEP-151][].
-
-### Details
-
-This rule validates that the schema for `202` responses includes the required
-`Operation` properties: `path`, `done`, `error`, and `response`.
-
-### Examples
-
-**Incorrect** code for this rule:
-
-```yaml
-paths:
-  '/books':
-    post:
-      summary: Create a book
-      responses:
-        '202':
-          description: Book creation accepted
-          content:
-            'application/json':
-              schema:
-                type: object
-                properties:
-                  path:
-                    type: string
-                  done:
-                    type: boolean
-                  # Missing error and response properties
-```
-
-**Correct** code for this rule:
-
-```yaml
-paths:
-  '/books':
-    post:
-      summary: Create a book
-      responses:
-        '202':
-          description: Book creation accepted
-          content:
-            'application/json':
-              schema:
-                type: object
-                properties:
-                  path:
-                    type: string
-                  done:
-                    type: boolean
-                  error:
-                    type: object
-                  response:
-                    type: object
-```
-
-### Disabling
-
-If you need to violate this rule for a specific operation, add an "override" to
-the Spectral rule file for the specific file and fragment.
-
-```yaml
-overrides:
-  - files:
-      - 'openapi.json#/paths/books/post/responses/202/content/application/json/schema'
-    rules:
-      aep-151-operation-schema: 'off'
-```
-
-## aep-151-operation-properties
 
 **Rule**: Operation schema properties must have correct types
